FXAA edge detection algorithm modified! SSAA Implemented!
* SSAA member added to Antialiasing enum.
* EdgeDetection enum class added.
* Camera::GenerateRay() signature changed. Row and Col are now floats instead of ints.
* ColorUtils ges a new getColorDifference() function, calculating Euclidean distance.
* New toFloat inline functino added to Render.cpp for quick and short unsigned->float casting.
* Render::RenderImage() and ::RnederTree() algorithm changed to account for SSAA.
* Settings:: subPixDepthAA and edgeDetectionTypeFXAA memebrs added.
* ray2NDC space conversion function signature changed to take row and column as floats instead of ints.
* ray2NDC no longer multiplies by 0.5. This logic was moved to the caller.
* FVector3::CalculateLength() now uses sqrtf instead of sqrt.

+++ RayTracer/inc/Camera.h
@@ -59,7 +59,7 @@ struct Camera : public Obj {
 	// @param[in] y: Resolution height.
 	// @param[in] settings: Scene settings.
 	// @return The generated camera ray.
-	Ray GenerateRay( const int, const int, const Settings& ) const;
+	Ray GenerateRay( const float, const float, const Settings& ) const;
 
 	// Move Camera forward-backward.
 	// @param[in] val: The number of units the Camera will move.
